Stop import cities

+++ uninstall.php
@@ -1,15 +1,17 @@
 <?php
+
     // if uninstall.php is not called by WordPress, die
     if ( ! defined('WP_UNINSTALL_PLUGIN' ) ) {
         die;
     }
 
     // drop the database table
-    global $wpdb;
-    $wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}cities");
+    if ( false == get_option( 'acfcs_preserve_settings' ) ) {
+	    global $wpdb;
+	    $wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}cities");
 
-	$target_folder = wp_upload_dir()['basedir'] . '/acfcs';
-    rmdir( $target_folder );
+	    $target_folder = wp_upload_dir()['basedir'] . '/acfcs';
+	    rmdir( $target_folder );
+    }
 
-    // drop all settings from db (if any)
 ?>
